refactor: interactive-elements
